Converted bucketUtil to an interface. Renamed variable and file names to be generic.
