+++ perfdash/Makefile
@@ -5,7 +5,7 @@ TAG = 2.25
 
 REPO = gcr.io/k8s-testimages
 
-test: perfdash.go parser.go config.go downloader.go google-gcs-downloader.go
+test: perfdash.go parser.go config.go downloader-helper.go downloader.go
 	go test
 
 perfdash: test
